+++ portfolio/portfolio/src/main/java/com/personal/portfolio/menu/MenuRepository.java
@@ -2,6 +2,6 @@ package com.personal.portfolio.menu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<MenuEntity, Integer> {
+public interface MenuRepository extends JpaRepository<MenuEntity, Long> {
 
 }
