+++ portfolio/portfolio/src/main/java/com/personal/portfolio/music/MusicRepository.java
@@ -2,5 +2,5 @@ package com.personal.portfolio.music;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MusicRepository extends JpaRepository<MusicEntity, Integer> {
+public interface MusicRepository extends JpaRepository<MusicEntity, Long> {
 }
